# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Deep Learning CNN+LSTM API",
    "description": "API REST pour le projet de Deep Learning avec CNN+LSTM",
    "version": "1.0.0",
    "contact": {
      "name": "Support",
      "email": "support@deeplearning.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Serveur de développement"
    }
  ],
...
  "paths": {
    "/api/model/create": {
      "post": {
        "tags": ["Model"],
        "summary": "Créer un nouveau modèle CNN+LSTM",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "inputShape": {
                    "type": "array",
                    "items": {"type": "integer"},
                    "example": [64, 64, 1],
                    "description": "Forme de l'entrée [hauteur, largeur, canaux]"
                  },
                  "numClasses": {
                    "type": "integer",
                    "example": 10,
                    "description": "Nombre de classes de sortie"
                  }
                },
                "required": ["inputShape", "numClasses"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Modèle créé avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {"type": "boolean"},
                    "message": {"type": "string"},
                    "modelSummary": {"type": "string"}
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erreur serveur"
          }
        }
      }
    },
...
    "/api/model/save": {
      "post": {
        "tags": ["Model"],
        "summary": "Sauvegarder le modèle",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "modelName": {
                    "type": "string",
                    "example": "mon-modele-v1"
                  }
                },
                "required": ["modelName"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Modèle sauvegardé"
          }
        }
      }
    },
    "/api/model/load": {
      "post": {
        "tags": ["Model"],
        "summary": "Charger un modèle",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "modelName": {
                    "type": "string",
                    "example": "mon-modele-v1"
                  }
                },
                "required": ["modelName"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Modèle chargé"
          }
        }
      }
    },
...
  },
...
}